support fs_game to specify gamedata dir start to split out game into dll move main menu into game dll hardcode console scene at the end

+++ src/scene.h
@@ -63,7 +63,7 @@ public:
 		newScene->Startup(&info);
 	}
 
-	Scene* Get(unsigned long i) {
+	Scene* Get(unsigned int i) {
 		if (i > scenes.size()) {
 			return nullptr;
 		}
